Removed AbstractDocumentType
