+++ src/Sulu/Bundle/ContentBundle/Form/Type/BasePageDocumentType.php
@@ -13,14 +13,56 @@ namespace Sulu\Bundle\ContentBundle\Form\Type;
 use Symfony\Component\Form\FormInterface;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
+use Sulu\Component\PHPCR\SessionManager\SessionManagerInterface;
+use Sulu\Component\DocumentManager\DocumentManager;
+use Symfony\Component\Form\FormEvents;
+use Symfony\Component\Form\FormEvent;
 
-abstract class BasePageDocumentType extends AbstractDocumentType
+abstract class BasePageDocumentType extends AbstractContentBehaviorType
 {
+    /**
+     * @var SessionManagerInterface
+     */
+    private $sessionManager;
+
+    /**
+     * @var DocumentManager
+     */
+    private $documentManager;
+
+    /**
+     * @param SessionManagerInterface $sessionManager
+     * @param DocumentManager $documentManager
+     */
+    public function __construct(
+        SessionManagerInterface $sessionManager,
+        DocumentManager $documentManager
+    )
+    {
+        $this->sessionManager = $sessionManager;
+        $this->documentManager = $documentManager;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function setDefaultOptions(OptionsResolverInterface $options)
+    {
+        parent::setDefaultOptions($options);
+        $options->setRequired(array(
+            'webspace_key',
+        ));
+    }
+
     /**
      * {@inheritDoc}
      */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
+        parent::buildForm($builder, $options);
+
+        $builder->add('parent', 'document_object');
+        $builder->add('extensions', 'text', array('property_path' => 'extensionsData'));
         $builder->add('resourceSegment', 'text');
         $builder->add('navigationContexts', 'collection', array(
             'type' => 'text',
@@ -33,7 +75,38 @@ abstract class BasePageDocumentType extends AbstractDocumentType
         $builder->add('workflowStage', 'integer');
         $builder->add('shadowLocaleEnabled', 'checkbox');
         $builder->add('shadowLocale', 'text'); // TODO: Should be choice of available shadow locales
+        $builder->setAttribute('webspace_key', $options['webspace_key']);
+        $builder->setAttribute('clear_missing_content', $options['clear_missing_content']);
 
-        parent::buildForm($builder, $options);
+        $builder->addEventListener(FormEvents::POST_SUBMIT, array($this, 'postSubmitDocumentParent'));
+    }
+
+    /**
+     * Set the document parent to be the webspace content path
+     * when the document has no parent.
+     *
+     * @param FormEvent $event
+     */
+    public function postSubmitDocumentParent(FormEvent $event)
+    {
+        $document = $event->getData();
+
+        if ($document->getParent()) {
+            return;
+        }
+
+        $form = $event->getForm();
+        $webspaceKey = $form->getConfig()->getAttribute('webspace_key');
+        $parent = $this->documentManager->find($this->sessionManager->getContentPath($webspaceKey));
+
+        if (null === $parent) {
+            throw new \InvalidArgumentException(sprintf(
+                'Could not determine parent for document with title "%s" in webspace "%s"',
+                $document->getTitle(),
+                $webspaceKey
+            ));
+        }
+
+        $document->setParent($parent);
     }
 }
