+++ src/Sulu/Component/Content/Mapper/ContentMapper.php
@@ -251,6 +251,7 @@ class ContentMapper implements ContentMapperInterface
         $shadowBaseLanguage = null,
         $documentAlias = LegacyStructure::TYPE_PAGE
     ) {
+        // map explicit arguments to data
         $data['parent'] = $parentUuid;
         $data['workflowStage'] = $state;
         $data['structureType'] = $structureType;
